backend/requirements.txt
# Python dependencies
fastapi==0.108.0
uvicorn[standard]==0.25.0
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
redis==5.0.1
pytest==7.4.4
pytest-asyncio==0.23.2
httpx==0.25.2
python-dotenv==1.0.0
aiosqlite==0.19.0  # For testing with SQLite
psutil==5.9.6  # For system performance monitoring

# Document processing and vector database
chromadb==0.4.22
sentence-transformers==2.2.2
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
tiktoken==0.5.2

# AI Provider SDKs
openai==1.6.1
anthropic==0.8.1
ollama==0.1.7
